Add incomplete solution to prop. prob. with missing input pattern
When finding a conditional propagation probability, you need to find
the intersection of propagation events. However, not every Event
object has the same input patterns. So, we need to test for the
case where an input pattern is missing. This fix only does it
for a single conditional probability. Further work is needed for
  multiple conditional probabilities.