Add option for buffer to be placed on specific branch
Previously, the bench.add_buffer method would add a
buffer on the entire signal line. However, we sometimes
only want it added on a single branch. This change adds
an optional second argument to the method that specifies
the output of the next gate on the branch the buffer should
be placed on.