Add tests for finding input patterns
This change made it so we only look at the locked file now instead
of the oracle. To find the input patterns, we not replace all key
gates with a buffer and put a fault at the output of the buffer.
This allows us to have faults at inputs and on a single branch
of a branching net, whereas before, when we were inserting faults
on the oracle, those cases would be incorrect.