Add package path to tests using conftest.py
I still have not figured out how to handle imports when running
pytest tests that work with imports when running python scripts
outside of pytest. However, this change allows imports to function
like if we were just running scripts and work when running tests.

It adds the ckttools directory to the path before any test is run.