Move first half of propagation tool to propagation namespace
It used to live in a top level tool of measure_propagation_events.py.
However, this isn't a tool that we will actually use. It will only
be integrated in the lambda tool. So, it makes sense to stick it
in a namespace.

The second half of the tool is the actual function that gets the
propagation events. It will be moved in the next change when there
are high level tests for it.