Don't remove primary inputs from bench file when finding input patterns
The way we find patterns is by removing all traces of a key gate
and replacing them with a buffer. This was done with a recursive
call to remove all gates associated with the key, including inputs.

However, some locking schemes include primary keys in the fanin of
a key gate, like SARLock. These were being removed. This change
makes it so those aren't removed when converting a verilog file
to a bench file.