Don't remove constants in recursive remove if they are still used
The remove_gates_recursive method gobbles up and gate it finds.
However, this causes an issue if there are constants being used
in multiple places of the circuit. This change makes it so the
remove_gate_recursive method checks to see if any constants are
still in use, and if not, it removes that circuitry.

Note that one could say "why don't we do this for every gate? After
all, any gate can be used more than once." To that we say, yes, but
the use case of this method should not include that. This should
only be used when you know you want to delete every gate. Constants
are an edge case, and supporting that seems reasonable.