Add option to limit how many faults Atalanta finds
With large circuits, the -A option for Atalanta finds all faults
and can take awhile. With c1908 (32 inputs), it took like 15ish
seconds. However, the -D option allows us to input how many faults
we want to find. This is almost a must use, as the c1908 found
1.5 million fault strings. Note that that is fault strings, each
of which have several unknown values.

This change gives an optional number of samples to the atalanta
runner. If it is present, it uses the -D flag. If not, it uses -A.