Add non-flipping key inputs to verilog metadata
Some keys don't flip the signal in the circuit. For example, adding
a multiplexer only switches between 2 signals, one of which has
been flipped and the other which hasn't been flipped. Measuring
lambda is only for the gates that have a chance to flip a signal
in the circuit.

This change adds new metadata to the circuit in the form of the
key inputs that do not flip a signal. These key inputs should
not be removed from the bench file because we need to essentially
consider them like they are a part of the oracle. So, it may be
better to say that the oracle is actually the circuit without
key gates that flip a signal.