Add lambda calculation for a single key gate
The current calculations are not correct. It is not calculating the
p(prop) correctly for sarlock circuits. It is calculating p(prop)
how we expect it to, but for some reason SARLock's p(prop) isn't
calculated like that. My notes say that we should omit the mask
part of SARLock and only look at the flip signal.

We need to adjust how we define one of the terms in lambda. Either
p(prop) cannot take into account the mask, or the number of incorrect
keys needs to be changed.