Add output-adjacent-non-unary filter
This filter finds the first net that is next the an output, where the
gate preceding the net must have more than 1 input. This is a very
clunky filter, and gives me an idea that there should be nested
filters somehow. For example, this one filter could be split into
output-adjacent and non-unary.