Complete SAT attack
This SAT attack is slightly different than the previous version and
causes some differences in the # iterations. Instead of creating the
output of each circuit into an entire z3 tree, each gate of the circuit
is added as a new constraint to the solver.

There is some work still needed. The new SAT attack is slower, and I
am unsure if it is caused by the SAT solving or the fact that we
iterate through each verilog AST twice to convert to z3.

Additionally, we should add both types of dip finders (legacy and new)
to be able to swap between them.