Add legacy dip finder to SAT attack
The legacy dip finder is much faster than the new version implemented.
I am not sure if this is because of the solver or because we don't
iteration through the whole AST twice. I still need to find out.

This change also creates a moddef wrapper which caches things about
the moddef so we don't have to iterate over the AST over and over.
This wrapper still needs to support modifying a moddef, as right now
it only supports reading from one.