Fix Antisat input bits to truly be continuous
Before, the "continuous" option when doing an Antisat block
would actually not be continuous - it would skip several input bits.
This was because it was offsetting by the number of key bits, but the
number of input bits is actually half of the number of key bits.

This change fixes that and also changes the IterationData class
to reflect the new change.