Add keycheck tool
This tool checks that the provided key is correct by using
a miter circuit with an unlocked version. It works when tested
with small circuits and a c5135 circuit with the provided key.
However, the key recovered from the SAT attack for the same c5135
circuit fails on this tool.

I am unsure if the issue is with this tool or with the SAT attack.