Convert z3 BoolRef to bools
There was an issue that z3's BoolRef type was not
being recognized as a bool. I did not notice this
because it is printed out exactly as a bool is. So,
this change casts those types to bools, allowing
our tools to not have to worry about the z3 reference.