Skip failing ChosenDipFinder test
This test was originally written when ChosenDipFinder would return
False from can_find_dip() when it ran out of chosen dips. However,
this behavior was not desired because I wanted to see what DIPs were
needed to complete the SAT attack after the DIPs I supplied ran out.

To accomplish this, the ChosenDipFinder hides a DefaultDipFinder that
it uses after the chosen dips run out. This test case asserts that
can_find_dip() returns False when DIPs run out. I am not sure how to
test that the implementation switches to a DefaultDipFinder without
fully setting up a SAT attack to run. For now, I'll skip this test
and probably never come back to it later.